plugins/ml2lm/requirements.txt
-i https://pypi.org/simple
amqp==2.3.2
autopep8==1.3.5
beautifulsoup4==4.6.0
billiard==3.5.0.3
bs4==0.0.1
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-material==1.2.5
django-rest-framework==0.1.0
django-shorturls==2.0.0
django-silk==3.0.0
django==2.0.6
djangorestframework==3.8.2
gprof2dot==2016.10.13
html5lib==1.0.1
idna==2.7
jinja2==2.10
kombu==4.2.1
markupsafe==1.0
psycopg2-binary==2.7.4
pycodestyle==2.4.0
pygments==2.2.0
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
vine==1.1.4
webencodings==0.5.1
